requirements.txt
appnope==0.1.2
backcall==0.2.0
blinker==1.4
certifi==2021.5.30
click==8.0.1
decorator==4.4.2
et-xmlfile==1.0.1
Flask==2.0.1
gunicorn==20.1.0
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openpyxl==3.0.7
parso==0.8.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
python-dotenv==0.19.0
selenium==3.141.0
sentry-sdk==1.3.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==2.0.1
requests==2.26.0
pytest==6.2.4
python-binance==1.0.12
